+++ src/app/components/map/KakaoMap.tsx
@@ -5,6 +5,7 @@ import { useEffect, useState } from "react";
 import { CustomOverlayMap, Map, MapMarker } from "react-kakao-maps-sdk";
 import ReSetttingMapBounds from "./ReSetttingMapBounds";
 import Link from "next/link";
+import RestaurantOverlay from "./RestaurantOverlay";
 
 type Props = {
   restaurants: Restaurant[];
@@ -39,9 +40,8 @@ export default function KakaoMap({ restaurants }: Props) {
         level={13}
       >
         {restaurants.map((rest, index) => (
-          <>
+          <div key={`marker__${rest.latitude}-${rest.longitude}-${index}`}>
             <MapMarker
-              key={`marker__${rest.latitude}-${rest.longitude}-${index}`}
               position={{
                 lat: Number(rest.latitude),
                 lng: Number(rest.longitude)
@@ -50,50 +50,9 @@ export default function KakaoMap({ restaurants }: Props) {
               onClick={() => setSelectedRestaurant(rest)}
             />
             {selectedRestaurant?.restaurant_name === rest.restaurant_name && (
-              <CustomOverlayMap
-                position={{
-                  lat: Number(rest.latitude),
-                  lng: Number(rest.longitude)
-                }}
-                yAnchor={1.4}
-              >
-                <div className="flex flex-col items-center bg-white border rounded-lg shadow-md p-2 w-[200px]">
-                  <div className="flex justify-between items-center w-full mb-2">
-                    <h2 className="text-[13px] font-bold">{rest.restaurant_name}</h2>
-                    <img
-                      alt="close"
-                      width="14"
-                      height="13"
-                      src="https://t1.daumcdn.net/localimg/localimages/07/mapjsapi/2x/bt_close.gif"
-                      className="cursor-pointer"
-                      onClick={() => setSelectedRestaurant(null)}
-                    />
-                  </div>
-                  <div>
-                    <div className="mb-2">
-                      <img
-                        src={rest.restaurant_img_url || ""}
-                        width="73"
-                        height="70"
-                        alt={rest.restaurant_name || "이미지 없음"}
-                      />
-                    </div>
-                    <div>
-                      <div className="flex gap-[10px] items-center">
-                        <p className="text-sm font-light">{rest.address?.split(" ").slice(0, 2).join(" ")}</p>
-                        <Link
-                          href={`https://map.kakao.com/link/search/${encodeURIComponent(rest.restaurant_name)}`}
-                          passHref
-                        >
-                          <span className="text-xs">상세보기</span>
-                        </Link>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </CustomOverlayMap>
+              <RestaurantOverlay rest={rest} setSelectedRestaurant={setSelectedRestaurant} />
             )}
-          </>
+          </div>
         ))}
         {points.length > 0 && <ReSetttingMapBounds points={points} />}
       </Map>
